+++ web/jobs/assets/pencil-DHqgRXr7.js
@@ -1,4 +1,4 @@
-import{c as a}from"./index-CZOVDDyb.js";/**
+import{c as a}from"./index-BLIKf1tA.js";/**
  * @license lucide-react v0.500.0 - ISC
  *
  * This source code is licensed under the ISC license.